+++ runtime/tests/nnfw_api/lib/fixtures.h
@@ -30,9 +30,9 @@
 inline uint64_t num_elems(const nnfw_tensorinfo *ti)
 {
   uint64_t n = 1;
-  for (uint32_t i = 0; i < ti->rank; ++i)
+  for (int32_t i = 0; i < ti->rank; ++i)
   {
-    n *= ti->dims[i];
+    n *= static_cast<uint64_t>(ti->dims[i]);
   }
   return n;
 }
